Move isPlainObject into its own file
